requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
tensorflow==2.13.0
Pillow==10.0.0
openCV-python==4.8.1
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.1.0
catboost==1.2.2